+++ auto/lib/Setup-Apps.ps1
@@ -4,6 +4,7 @@ $scriptsLib = [IO.Path]::GetDirectoryName($MyInvocation.MyCommand.Definition)
 . "$scriptsLib\bench.lib.ps1"
 . "$scriptsLib\appconfig.lib.ps1"
 . "$scriptsLib\env.lib.ps1"
+. "$scriptsLib\adornment.lib.ps1"
 . "$scriptsLib\launcher.lib.ps1"
 
 Set-Debugging $debug
@@ -111,6 +112,7 @@ function Setup-Common([string]$name) {
     Load-AppEnvironment $name
     Execute-AppCustomSetup $name
     Execute-AppEnvironmentSetup $name
+    Setup-ExecutionProxies $name
     Create-Launcher $name
 }